LimApi/requirements.txt
pyinstrument==4.4.0
mysqlclient==2.1.1
pymysql==1.0.2
django==4.1.3
djangorestframework==3.14.0
django-cors-headers==3.13.0
django-filter==22.1
requests==2.28.1
sshtunnel==0.4.0
redis==4.4.0
openpyxl==3.1.0